Cooling Thermal Hotspots
Heat Transfer course project: This was an open-ended project where we had to take up a practical problem related to heat transfer and find an innovative solution for the same. Our group has solved the problem of thermal hotspots by cooling using sound waves. Details can be found in the report.